Move bloom logic to bloom.rs
